chore: fix typo `Algorithme` to  `Algorithm`
