+++ src/main/java/fr/iut/pathpilotapi/algorithm/AlgorithmType.java
@@ -1,27 +1,27 @@
-package fr.iut.pathpilotapi.algorithme;
+package fr.iut.pathpilotapi.algorithm;
 
 import lombok.Getter;
 
 /**
  * Enum to represent the different types of algorithms to find the best path to optimize the itinerary.
  */
-public enum AlgorithmeType {
+public enum AlgorithmType {
     BRUTE_FORCE(BruteForce.class),
     BRUTE_FORCE_THREAD(BruteForceThread.class),
     BRANCH_AND_BOUND(BranchAndBound.class);
 
     @Getter
     private final String name;
-    private final Class<? extends Algorithme> algorithme;
+    private final Class<? extends Algorithm> algorithm;
 
-    AlgorithmeType(Class<? extends Algorithme> algorithme) {
-        this.name = algorithme.getSimpleName();
-        this.algorithme = algorithme;
+    AlgorithmType(Class<? extends Algorithm> algorithm) {
+        this.name = algorithm.getSimpleName();
+        this.algorithm = algorithm;
     }
 
-    public Algorithme newInstanceAlgorithm() {
+    public Algorithm newInstanceAlgorithm() {
         try {
-            return algorithme.getDeclaredConstructor().newInstance();
+            return algorithm.getDeclaredConstructor().newInstance();
         } catch (Exception e) {
             throw new RuntimeException("Error while creating the instance of the algorithm", e);
         }
